change channel handling so that it's client side only
before the day parameter was first sent to server,
then server sent it back with the channel json list
so for each day the json list had to be fetched from
server.

this was "unoptimal", so now channel list is fetched
only once, is cached, and the day parameter is
"injected" with javascript.