fix race condition with ajax requests
If an ajax call took long, and a new ajax was initiated,
it could happen so that the new one finishes before the old.
Then the old request response handler overwrites the results
new had request already delivered.

This is fixed by aborting the old request when making the new one.
This required a "bleeding edge" (not year officially released")
version of core-ajax, where an abort method is added, it probably
will be released officially soon. The relevant commit is
https://github.com/Polymer/core-ajax/commit/f69f7cbb872319a48cb4ac133c518687193576b2
which I added to my 0.5.3 version of core-ajax.