Change video play to be "inline" in program list
This means the url and the list of programs is not changed when a title is
clicked, rather the program list functions as a dropdown menu, like tvkaista
own mobile web interface. This makes browsing programs and viewing descriptions
fast. Previously pressing back was both slow and lost the position of the list.

Also add thumbnail pictures to video placeholder ("poster").