test.py: return 500 items from search "elokuva"
This is because it's close to the number of movies currently
is in tvkaista 4 weeks buffer. Also it demonstrates the
problem with the results list - try this test.py server
with an older versions of the app, commits

f93fc5fdc94ee061f42df5a96869e334b67767af

where performance was just slow and

7a663182710f5a933e096a860b858a47b54387a

where the page never rendered on mobile... These problems
were before this only visible in production, with larger lists.